Org cascade destroy fix for org without owner

+++ app/models/organization.rb
@@ -59,7 +59,7 @@ class Organization < Sequel::Model
   def destroy_cascade
     destroy_permissions
     destroy_non_owner_users
-    self.owner.destroy
+    self.owner.destroy if self.owner
   end
 
   def destroy_permissions
